DCDO-3: security group module
The security group module needs to following parameters:

  * vpc_id
  * sg_name: Security group name. tags owner and environment
    will be appended
  * sg_description: Description
  * ingress_rules_cidr: A list of maps with cidr blocks as
    source
  * ingress_rules_sgid: A list of maps with security groups
    as source

`ingress_rules_cidr` and `ingress_rules_sgid` differ in the
type of source. The first one takes a comma separated list of
cidr blocks, and the second one, security groups.

An example of the `ingress_rules_cidr` is:
```
[
  {
    protocol    = "tcp",
    from_port   = "80",
    to_port     = "80",
    desc        = "HTTP port"
    cidr_blocks = "10.1.2.3/32, 10.2.3.4/32"
  }
]
```

The structure for `ingress_rules_sgid` is the same, with the
exception of the `cidr_blocks` attribute, which is replaced by
`sg_ids`