COR-373: Add EBS volumes to master and slave
Added an additional EBS volume to master and slave on
`launch_block_device_mappings`. This device/s need to be named the
same way as keys in the ansible var `common_disk`.

FIXME:
```
"{{ user `MASTER_XVDF_SIZE`}}"
```
Does not work, as it needs to be an integer, not a string.
Removing the double quotes only generates an error.
(same applies for `SLAVE_...`)